requirements.txt
playwright==1.40.0
psycopg2-binary>=2.9.0
python-dotenv==1.0.0
lxml==4.9.3
requests==2.31.0
click==8.1.7
schedule==1.2.0
flask>=2.0.0

